requirements.txt
requests==2.31.0
psycopg2-binary==2.9.7
pydantic==2.4.2
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.24.3